[#4] refactor: conflict 없도록 프로젝트 게시 API 수정

+++ src/main/java/com/artistack/instrument/repository/InstrumentRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface InstrumentRepository extends JpaRepository<Instrument, Long> {
-    Optional<Instrument> findById(Integer id);
+    Optional<Instrument> findById(Long id);
 }
